remove support for floats
Because .9999999999998 yk
I ain't making a decision on what to do with that situation, because
rounding to *any* point is imprecise, and showing the actual number is
wrong, so this isn't worth it

+++ src/actions.rs
@@ -12,18 +12,18 @@ pub fn clear() -> Result<(), io::Error> {
 }
 
 pub fn increment(counter: PathBuf) -> Result<(), Box<dyn Error>> {
-    data::read_write(&counter, |value| value + 1.0)
+    data::read_write(&counter, |value| value + 1)
 }
 
 pub fn decrement(counter: PathBuf) -> Result<(), Box<dyn Error>> {
-    data::read_write(&counter, |value| value - 1.0)
+    data::read_write(&counter, |value| value - 1)
 }
 
-pub fn set(counter: PathBuf, num: f64) -> Result<(), Box<dyn Error>> {
+pub fn set(counter: PathBuf, num: i64) -> Result<(), Box<dyn Error>> {
     Ok(data::write(&counter, num)?)
 }
 
-pub fn add(counter: PathBuf, num: f64) -> Result<(), Box<dyn Error>> {
+pub fn add(counter: PathBuf, num: i64) -> Result<(), Box<dyn Error>> {
     data::read_write(&counter, |value| value + num)
 }
